Refactor missed actions area to add room for more items in the future
