+++ src/Patient/Home/HomePage.js
@@ -3,13 +3,13 @@ import Progress from './Progress';
 import ActionBox from './ActionBox';
 import Alerts from './Alerts';
 import Colors from '../../Basics/Colors';
-import Greeting from './Greeting'
-import Reminders from './Reminders'
-import MedicationReminder from './MedicationReminder'
-import CachedReports from './CachedReports'
-import VideoCard from './Videos'
-import MissedReports from './MissedDays'
-import PushNotificationEnrollment from './PushEnrollmentReminder'
+import Greeting from './Greeting';
+import Reminders from './Reminders';
+import MedicationReminder from './MedicationReminder';
+import CachedReports from './CachedReports';
+import VideoCard from './Videos';
+import RequiresAction from './MissedActions';
+import PushNotificationEnrollment from './PushEnrollmentReminder';
 import { makeStyles } from '@material-ui/core/styles';
 
 const useStyles = makeStyles({
@@ -32,7 +32,7 @@ const HomePage = () => {
             <PushNotificationEnrollment />
             <Greeting />
             <ActionBox />
-            <MissedReports />
+            <RequiresAction />
             <VideoCard />
             <CachedReports />
             <Alerts />
